fix :  optimize page load times with ISR, caching, and query optimization
