+++ next.config.mjs
@@ -4,6 +4,14 @@ const nextConfig = {
   swcMinify: true,
   images: {
     domains: ["res.cloudinary.com", "images.unsplash.com"],
+    formats: ['image/avif', 'image/webp'],
+    minimumCacheTTL: 60,
+  },
+  compress: true,
+  poweredByHeader: false,
+  generateEtags: true,
+  experimental: {
+    optimizePackageImports: ['lucide-react', '@/components/ui'],
   },
 };
