+++ src/app/(main)/(pages)/about/about-us/page.tsx
@@ -7,6 +7,9 @@ export const metadata = {
     "Learn about BUCC's mission, vision, and journey since 2001. Discover how we empower tech enthusiasts and shape future leaders in technology.",
 };
 
+// Enable ISR with 120 second revalidation (events change less frequently here)
+export const revalidate = 120;
+
 export default async function AboutUs() {
   const eventsResult = await getAllEvents();
   const allEvents = eventsResult.success ? eventsResult.data ?? [] : [];
